+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Added [`search()`](./docs/jobs#search) function to Jobs class
+
 # 0.4.0 (2019-12-18)
 
 ### Added